Move JumpToLastMessageButton into a separate file
